python-app/requirements.txt
paho-mqtt==1.5.1
Pillow==7.2.0